### 2017-07-16 #318 @xuanxu
@arfon I think the review should be easy to someone with working knowledge of Ruby. The "genomics" part of the software refers to the values of the expected input data, but there's no need to be a genomics expert to test the software. After reading the lists of editors & reviewers I think you or @amoeba are good options. But if you prefer I can ask a colleague for help.
