### 2017-08-04 #318 @fstrozzi
Sure I'd be glad to. I'll need few days though, to have a bit of time to look at it.
