### 2017-08-04 #351 @whedon
[REVIEW]: Nimbus: a Ruby gem to implement Random Forest algorithms in a genomic selection context
**Submitting author:** @xuanxu (<a href="http://orcid.org/0000-0001-7699-3983">Juanjo Bazán</a>)
**Repository:** <a href="https://github.com/xuanxu/nimbus" target ="_blank">https://github.com/xuanxu/nimbus</a>
**Version:** 2.3.0
