### 2017-08-12 #351 @fstrozzi
The authors developed a Ruby package enabling random forest based classification for genomics selection data. The package is well written, quite easy to use and overall it seems a solid piece of software addressing specific analysis using SNPs data. The package is answering all JOSS requirements, although minor revisions are envisaged (marked as checkboxes below).

## Code

* the code is clean, and commented.
* rdoc is available online
* installation of the software works smoothly using RubyGem and there are no dependencies issues that I can envisage
* rspec automated tests are provided and the repository is linked to Travis CI

## Functionalities

* examples works fine with no errors following Readme instructions and using the example data provided into the repository

- [ ]  automated tests fail, an [issue](https://github.com/xuanxu/nimbus/issues/4) has been opened on the repository to inspect this

## Paper

- [ ]  The paper describes clearly the background of random forests algorithm and the use of Nimbus, but the statements of need could be improved by describing what Nimbus adds to existing random forest classifiers available for genomics selection, and highlighting what are the strong points of Nimbus and the reason why it was developed (e.g. simplicity to use, better fit for specific use cases, better performance etc.)

- [ ] Minor grammar and typing errors (can be easily spotted using a word processor, Markdown isn't very helpful with this)
