### 2017-08-12 #351 @pjotrp
@fstrozzi thanks for your thorough review! @xuanxu looks like these are minor fixes. Ping us when you are done or respond here.
