### 2017-08-16 xuanxu/nimbus#4 @xuanxu
Thanks @fstrozzi
I can not reproduce the fails. All tests pass in my local machine (Mac OSX 10.11.6 using Ruby 2.4.1), and in the [Travis server builds](https://travis-ci.org/xuanxu/nimbus). I'm using ruby 2.4.1 instead of 2.4.0 but I don't think that is the cause of the errors.

Maybe is the command used to run the specs? could you try `bundle exec rake` or `bundle exec rspec`instead of just `rake`?
