### 2017-08-16 449d60c
fixes grammar and typing errors in MD files
