### 2017-08-16 xuanxu/nimbus#4 @fstrozzi
Hi @xuanxu,
even with bundle exec I do have the same tests failing. I tried installing Ruby 2.4.1 and got the same result (below). Not a big issue if everything else on Travis is running fine, but still looks a bit weird since Nimbus then run absolutely fine with no errors.

```
> ruby -v
ruby 2.4.1p111 (2017-03-22 revision 58053) [x86_64-darwin15]

>bundle install
Fetching gem metadata from https://rubygems.org/..........
Fetching version metadata from https://rubygems.org/.
Using rake 12.0.0
Using bundler 1.15.3
Fetching diff-lcs 1.3
Installing diff-lcs 1.3
Using nimbus 2.3.0 from source at `.`
Fetching rspec-support 3.6.0
Installing rspec-support 3.6.0
Fetching rspec-core 3.6.0
Installing rspec-core 3.6.0
Fetching rspec-expectations 3.6.0
Installing rspec-expectations 3.6.0
Fetching rspec-mocks 3.6.0
Installing rspec-mocks 3.6.0
Fetching rspec 3.6.0
Installing rspec 3.6.0
Bundle complete! 3 Gemfile dependencies, 9 gems now installed.
Use `bundle info [gemname]` to see where a bundled gem is installed.

>bundle exec rake
/Users/fstrozzi/.rbenv/versions/2.4.1/bin/ruby -I/Users/fstrozzi/.rbenv/versions/2.4.1/lib/ruby/gems/2.4.0/gems/rspec-core-3.6.0/lib:/Users/fstrozzi/.rbenv/versions/2.4.1/lib/ruby/gems/2.4.0/gems/rspec-support-3.6.0/lib /Users/fstrozzi/.rbenv/versions/2.4.1/lib/ruby/gems/2.4.0/gems/rspec-core-3.6.0/exe/rspec --pattern spec/\*\*\{,/\*/\*\*\}/\*_spec.rb

Nimbus::ClassificationTree
  is initialized with tree config info
  creates a tree structure when seeded with training data
  splits node when building a node and finds a suitable split
  keeps track of all SNPs used for the tree
  labels node when building a node and there is not a suitable split
  labels node when building a node with less individuals than the minimum node size
  computes generalization error for the tree
  estimates importance for all SNPs
  get prediction for an individual pushing it down a tree structure (FAILED - 1)

Nimbus::Configuration
  loads configuration options from config file (FAILED - 2)
  tree method return tree-related subset of options for regression trees
  tree method return tree-related subset of options for classification trees
  creates a training set object from training data file (FAILED - 3)
  reads testing data and yields one individual at a time
  creates a forest object loading data from a yaml file (FAILED - 4)

Nimbus::Forest
  Regression
Creating trees: 1/3     grows a regression forest of N trees (FAILED - 5)
Creating trees: 1/3     creates averaged predictions for individuals in the training set (FAILED - 6)
Creating trees: 1/3     computes averaged SNP importances for every SNP (FAILED - 7)
Creating trees: 1/3     does not compute SNP importances if config set to false (FAILED - 8)
    traverses a set of testing individuals through every tree in the forest and returns predictions
    can output forest structure in YAML format
  Classification
    grows a classification forest of N trees
    creates predictions for individuals in the training set
    computes averaged SNP importances for every SNP
    does not compute SNP importances if config set to false
    traverses a set of testing individuals through every tree in the forest and returns predictions
    can output forest structure in YAML format

Nimbus::Individual
  stores id, fenotype and SNPs information for an individual

Nimbus::LossFunctions
  method for average
  method for mean squared error
  method for quadratic_loss
  quadratic loss is mean squared error averaged
  method for pseudo Huber error
  method for pseudo Huber loss function
  pseudo Huber loss is pseudo Huber error averaged
  method for squared difference
  method for majority class
  majority class method selects randomly if more than one majority class
  method for majority class in list
  method for class sizes
  method for class sizes in list
  Gini index

Nimbus module
  manages a Nimbus::Application object
  accepts setting an external Nimbus::Application

Nimbus::RegressionTree
  is initialized with tree config info
  creates a tree structure when seeded with training data (FAILED - 9)
  split node when building a node and finds a suitable split (FAILED - 10)
  keeps track of all SNPs used for the tree (FAILED - 11)
  labels node when building a node and there is not a suitable split (FAILED - 12)
  labels node when building a node with less individuals than the minimum node size (FAILED - 13)
  computes generalization error for the tree (FAILED - 14)
  estimates importance for all SNPs (FAILED - 15)
  get prediction for an individual pushing it down a tree structure (FAILED - 16)

Nimbus::TrainingSet
  stores individuals list and fenotype data for them
  keeps track of ids of all individuals in the training set

Nimbus::Tree
  is initialized with tree config info

Failures:

  1) Nimbus::ClassificationTree get prediction for an individual pushing it down a tree structure
     Failure/Error: expect(prediction).to eq '0'

       expected: "0"
            got: "1"

       (compared using ==)
     # ./spec/classification_tree_spec.rb:122:in `block (2 levels) in <top (required)>'

  2) Nimbus::Configuration loads configuration options from config file
     Failure/Error: expect(config.training_file).to eq fixture_file('regression/training.data')

       expected: "/Users/fstrozzi/dev/joss_review/nimbus/spec/fixtures/regression/training.data"
            got: nil

       (compared using ==)
     # ./spec/configuration_spec.rb:10:in `block (2 levels) in <top (required)>'

  3) Nimbus::Configuration creates a training set object from training data file
     Failure/Error:
             File.open(@training_file) {|file|
               @training_set = Nimbus::TrainingSet.new({}, {})
               file.each do |line|
                 next if line.strip == ''
                 data_feno, data_id, *snp_list = line.strip.split
                 raise Nimbus::InputFileError, "Individual ##{data_id} from training set has no value for all #{@tree_SNP_total_count} SNPs" unless snp_list.size == @tree_SNP_total_count
                 raise Nimbus::InputFileError, "There are individuals with no ID, please check data in training file." unless (!data_id.nil? && data_id.strip != '')
                 raise Nimbus::InputFileError, "Individual ##{data_id} has no fenotype value, please check data in training file." unless (!data_feno.nil? && data_feno.strip != '')
                 raise Nimbus::InputFileError, "Individual ##{data_id} has invalid class (not in [#{classes*', '}]), please check data in training file." unless (@classes.nil? || @classes.include?(data_feno))

     TypeError:
       no implicit conversion of nil into String
     # ./lib/nimbus/configuration.rb:153:in `initialize'
     # ./lib/nimbus/configuration.rb:153:in `open'
     # ./lib/nimbus/configuration.rb:153:in `load_training_data'
     # ./spec/configuration_spec.rb:62:in `block (2 levels) in <top (required)>'

  4) Nimbus::Configuration creates a forest object loading data from a yaml file
     Failure/Error: expect(trees.first.keys.first).to eq 176

       expected: 176
            got: 149

       (compared using ==)
     # ./spec/configuration_spec.rb:111:in `block (2 levels) in <top (required)>'

  5) Nimbus::Forest Regression grows a regression forest of N trees
     Failure/Error: @bag ||= @options.training_set.all_ids

     NoMethodError:
       undefined method `all_ids' for nil:NilClass
     # ./lib/nimbus/forest.rb:110:in `bag'
     # ./lib/nimbus/forest.rb:102:in `individuals_random_sample'
     # ./lib/nimbus/forest.rb:39:in `block in grow'
     # ./lib/nimbus/forest.rb:37:in `times'
     # ./lib/nimbus/forest.rb:37:in `grow'
     # ./spec/forest_spec.rb:18:in `block (3 levels) in <top (required)>'

  6) Nimbus::Forest Regression creates averaged predictions for individuals in the training set
     Failure/Error: @bag ||= @options.training_set.all_ids

     NoMethodError:
       undefined method `all_ids' for nil:NilClass
     # ./lib/nimbus/forest.rb:110:in `bag'
     # ./lib/nimbus/forest.rb:102:in `individuals_random_sample'
     # ./lib/nimbus/forest.rb:39:in `block in grow'
     # ./lib/nimbus/forest.rb:37:in `times'
     # ./lib/nimbus/forest.rb:37:in `grow'
     # ./spec/forest_spec.rb:25:in `block (3 levels) in <top (required)>'

  7) Nimbus::Forest Regression computes averaged SNP importances for every SNP
     Failure/Error: @bag ||= @options.training_set.all_ids

     NoMethodError:
       undefined method `all_ids' for nil:NilClass
     # ./lib/nimbus/forest.rb:110:in `bag'
     # ./lib/nimbus/forest.rb:102:in `individuals_random_sample'
     # ./lib/nimbus/forest.rb:39:in `block in grow'
     # ./lib/nimbus/forest.rb:37:in `times'
     # ./lib/nimbus/forest.rb:37:in `grow'
     # ./spec/forest_spec.rb:32:in `block (3 levels) in <top (required)>'

  8) Nimbus::Forest Regression does not compute SNP importances if config set to false
     Failure/Error: @bag ||= @options.training_set.all_ids

     NoMethodError:
       undefined method `all_ids' for nil:NilClass
     # ./lib/nimbus/forest.rb:110:in `bag'
     # ./lib/nimbus/forest.rb:102:in `individuals_random_sample'
     # ./lib/nimbus/forest.rb:39:in `block in grow'
     # ./lib/nimbus/forest.rb:37:in `times'
     # ./lib/nimbus/forest.rb:37:in `grow'
     # ./spec/forest_spec.rb:40:in `block (3 levels) in <top (required)>'

  9) Nimbus::RegressionTree creates a tree structure when seeded with training data
     Failure/Error:
             File.open(@training_file) {|file|
               @training_set = Nimbus::TrainingSet.new({}, {})
               file.each do |line|
                 next if line.strip == ''
                 data_feno, data_id, *snp_list = line.strip.split
                 raise Nimbus::InputFileError, "Individual ##{data_id} from training set has no value for all #{@tree_SNP_total_count} SNPs" unless snp_list.size == @tree_SNP_total_count
                 raise Nimbus::InputFileError, "There are individuals with no ID, please check data in training file." unless (!data_id.nil? && data_id.strip != '')
                 raise Nimbus::InputFileError, "Individual ##{data_id} has no fenotype value, please check data in training file." unless (!data_feno.nil? && data_feno.strip != '')
                 raise Nimbus::InputFileError, "Individual ##{data_id} has invalid class (not in [#{classes*', '}]), please check data in training file." unless (@classes.nil? || @classes.include?(data_feno))

     TypeError:
       no implicit conversion of nil into String
     # ./lib/nimbus/configuration.rb:153:in `initialize'
     # ./lib/nimbus/configuration.rb:153:in `open'
     # ./lib/nimbus/configuration.rb:153:in `load_training_data'
     # ./spec/regression_tree_spec.rb:19:in `block (2 levels) in <top (required)>'

  10) Nimbus::RegressionTree split node when building a node and finds a suitable split
      Failure/Error:
              File.open(@training_file) {|file|
                @training_set = Nimbus::TrainingSet.new({}, {})
                file.each do |line|
                  next if line.strip == ''
                  data_feno, data_id, *snp_list = line.strip.split
                  raise Nimbus::InputFileError, "Individual ##{data_id} from training set has no value for all #{@tree_SNP_total_count} SNPs" unless snp_list.size == @tree_SNP_total_count
                  raise Nimbus::InputFileError, "There are individuals with no ID, please check data in training file." unless (!data_id.nil? && data_id.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has no fenotype value, please check data in training file." unless (!data_feno.nil? && data_feno.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has invalid class (not in [#{classes*', '}]), please check data in training file." unless (@classes.nil? || @classes.include?(data_feno))

      TypeError:
        no implicit conversion of nil into String
      # ./lib/nimbus/configuration.rb:153:in `initialize'
      # ./lib/nimbus/configuration.rb:153:in `open'
      # ./lib/nimbus/configuration.rb:153:in `load_training_data'
      # ./spec/regression_tree_spec.rb:30:in `block (2 levels) in <top (required)>'

  11) Nimbus::RegressionTree keeps track of all SNPs used for the tree
      Failure/Error:
              File.open(@training_file) {|file|
                @training_set = Nimbus::TrainingSet.new({}, {})
                file.each do |line|
                  next if line.strip == ''
                  data_feno, data_id, *snp_list = line.strip.split
                  raise Nimbus::InputFileError, "Individual ##{data_id} from training set has no value for all #{@tree_SNP_total_count} SNPs" unless snp_list.size == @tree_SNP_total_count
                  raise Nimbus::InputFileError, "There are individuals with no ID, please check data in training file." unless (!data_id.nil? && data_id.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has no fenotype value, please check data in training file." unless (!data_feno.nil? && data_feno.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has invalid class (not in [#{classes*', '}]), please check data in training file." unless (@classes.nil? || @classes.include?(data_feno))

      TypeError:
        no implicit conversion of nil into String
      # ./lib/nimbus/configuration.rb:153:in `initialize'
      # ./lib/nimbus/configuration.rb:153:in `open'
      # ./lib/nimbus/configuration.rb:153:in `load_training_data'
      # ./spec/regression_tree_spec.rb:48:in `block (2 levels) in <top (required)>'

  12) Nimbus::RegressionTree labels node when building a node and there is not a suitable split
      Failure/Error:
              File.open(@training_file) {|file|
                @training_set = Nimbus::TrainingSet.new({}, {})
                file.each do |line|
                  next if line.strip == ''
                  data_feno, data_id, *snp_list = line.strip.split
                  raise Nimbus::InputFileError, "Individual ##{data_id} from training set has no value for all #{@tree_SNP_total_count} SNPs" unless snp_list.size == @tree_SNP_total_count
                  raise Nimbus::InputFileError, "There are individuals with no ID, please check data in training file." unless (!data_id.nil? && data_id.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has no fenotype value, please check data in training file." unless (!data_feno.nil? && data_feno.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has invalid class (not in [#{classes*', '}]), please check data in training file." unless (@classes.nil? || @classes.include?(data_feno))

      TypeError:
        no implicit conversion of nil into String
      # ./lib/nimbus/configuration.rb:153:in `initialize'
      # ./lib/nimbus/configuration.rb:153:in `open'
      # ./lib/nimbus/configuration.rb:153:in `load_training_data'
      # ./spec/regression_tree_spec.rb:60:in `block (2 levels) in <top (required)>'

  13) Nimbus::RegressionTree labels node when building a node with less individuals than the minimum node size
      Failure/Error:
              File.open(@training_file) {|file|
                @training_set = Nimbus::TrainingSet.new({}, {})
                file.each do |line|
                  next if line.strip == ''
                  data_feno, data_id, *snp_list = line.strip.split
                  raise Nimbus::InputFileError, "Individual ##{data_id} from training set has no value for all #{@tree_SNP_total_count} SNPs" unless snp_list.size == @tree_SNP_total_count
                  raise Nimbus::InputFileError, "There are individuals with no ID, please check data in training file." unless (!data_id.nil? && data_id.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has no fenotype value, please check data in training file." unless (!data_feno.nil? && data_feno.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has invalid class (not in [#{classes*', '}]), please check data in training file." unless (@classes.nil? || @classes.include?(data_feno))

      TypeError:
        no implicit conversion of nil into String
      # ./lib/nimbus/configuration.rb:153:in `initialize'
      # ./lib/nimbus/configuration.rb:153:in `open'
      # ./lib/nimbus/configuration.rb:153:in `load_training_data'
      # ./spec/regression_tree_spec.rb:75:in `block (2 levels) in <top (required)>'

  14) Nimbus::RegressionTree computes generalization error for the tree
      Failure/Error:
              File.open(@training_file) {|file|
                @training_set = Nimbus::TrainingSet.new({}, {})
                file.each do |line|
                  next if line.strip == ''
                  data_feno, data_id, *snp_list = line.strip.split
                  raise Nimbus::InputFileError, "Individual ##{data_id} from training set has no value for all #{@tree_SNP_total_count} SNPs" unless snp_list.size == @tree_SNP_total_count
                  raise Nimbus::InputFileError, "There are individuals with no ID, please check data in training file." unless (!data_id.nil? && data_id.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has no fenotype value, please check data in training file." unless (!data_feno.nil? && data_feno.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has invalid class (not in [#{classes*', '}]), please check data in training file." unless (@classes.nil? || @classes.include?(data_feno))

      TypeError:
        no implicit conversion of nil into String
      # ./lib/nimbus/configuration.rb:153:in `initialize'
      # ./lib/nimbus/configuration.rb:153:in `open'
      # ./lib/nimbus/configuration.rb:153:in `load_training_data'
      # ./spec/regression_tree_spec.rb:96:in `block (2 levels) in <top (required)>'

  15) Nimbus::RegressionTree estimates importance for all SNPs
      Failure/Error:
              File.open(@training_file) {|file|
                @training_set = Nimbus::TrainingSet.new({}, {})
                file.each do |line|
                  next if line.strip == ''
                  data_feno, data_id, *snp_list = line.strip.split
                  raise Nimbus::InputFileError, "Individual ##{data_id} from training set has no value for all #{@tree_SNP_total_count} SNPs" unless snp_list.size == @tree_SNP_total_count
                  raise Nimbus::InputFileError, "There are individuals with no ID, please check data in training file." unless (!data_id.nil? && data_id.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has no fenotype value, please check data in training file." unless (!data_feno.nil? && data_feno.strip != '')
                  raise Nimbus::InputFileError, "Individual ##{data_id} has invalid class (not in [#{classes*', '}]), please check data in training file." unless (@classes.nil? || @classes.include?(data_feno))

      TypeError:
        no implicit conversion of nil into String
      # ./lib/nimbus/configuration.rb:153:in `initialize'
      # ./lib/nimbus/configuration.rb:153:in `open'
      # ./lib/nimbus/configuration.rb:153:in `load_training_data'
      # ./spec/regression_tree_spec.rb:106:in `block (2 levels) in <top (required)>'

  16) Nimbus::RegressionTree get prediction for an individual pushing it down a tree structure
      Failure/Error: expect(prediction).to eq -0.90813

        expected: -0.90813
             got: -0.71043

        (compared using ==)
      # ./spec/regression_tree_spec.rb:119:in `block (2 levels) in <top (required)>'

Finished in 7.08 seconds (files took 0.46704 seconds to load)
56 examples, 16 failures

Failed examples:

rspec ./spec/classification_tree_spec.rb:118 # Nimbus::ClassificationTree get prediction for an individual pushing it down a tree structure
[…]
