### 2017-08-16 xuanxu/nimbus#4 @xuanxu
Hi @fstrozzi, yeah, it looks weird, the error messages suggest that rspec is not finding the right data files. Have you moved/change/delete any .data or config file? Or maybe add one to the base dir?
I'm still unable to reproduce the errors :S
