### 2017-08-16 #351 @xuanxu
Thanks for the review @fstrozzi!
I've add some details to the paper summary to highlight the strong points and improve the statement of need. And also corrected some typos.
