### 2017-08-17 xuanxu/nimbus#4 @fstrozzi
Hi @xuanxu,
the config or data files were untouched, the only thing that was interfering with the tests were the *.txt output files into the fixtures folders that were generated following the examples for the classification and regression tasks. Deleting those resolved the problem, so all green now and I am closing the issue.
