### 2017-08-17 #351 @fstrozzi
Hi @xuanxu,
the new version of the paper.md looks good (only a couple of typos left, in the second paragraph "accomodate" should be "accommodate" and in the very last line "use" should be "used"). I have also closed the issue on the repository since after the last exchanges we had, the tests run fine also on my end now.

@pjotrp from my point of view all the points raised were addressed.
