### 2017-08-17 e279cb0
fixes typos
