### 2017-08-17 #351 @xuanxu
Nice!
Both typos fixed.
